chore(alpha0): clean up code base before alpha 0 - INNO-1239
