+++ package.json
@@ -85,7 +85,6 @@
   "workspaces": [
     "examples/*",
     "src/systems/generic/design-tokens",
-    "src/systems/generic/implementations/react/*",
     "src/systems/generic/implementations/vanilla/*",
     "src/systems/generic/resources",
     "src/systems/ec/design-tokens",
