+++ src/systems/ec/implementations/react/packages/text-area/package.json
@@ -23,15 +23,13 @@
     "design-system",
     "react"
   ],
-  "dependencies": {
-    "@ecl/generic-react-component-text-area": "^2.0.0-alpha.0"
-  },
   "devDependencies": {
     "@ecl/ec-component-form-text-area": "^2.0.0-alpha.0",
     "@ecl/ec-specs-text-area": "^2.0.0-alpha.0",
     "@ecl/ec-theme-default": "^2.0.0-alpha.0"
   },
   "peerDependencies": {
+    "classnames": "^2.2.6",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
     "prop-types": "^15.6.2"
