+++ src/systems/eu/implementations/react/packages/blockquote/package.json
@@ -23,13 +23,10 @@
   ],
   "devDependencies": {
     "@ecl/eu-component-blockquote": "^2.0.0-alpha.0",
-    "@ecl/eu-specs-blockquote": "^2.0.0-alpha.0",
-    "@ecl/generic-react-component-blockquote": "^2.0.0-alpha.0",
-    "prop-types": "^15.6.2",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2"
+    "@ecl/eu-specs-blockquote": "^2.0.0-alpha.0"
   },
   "peerDependencies": {
+    "classnames": "^2.2.6",
     "prop-types": "^15.6.2",
     "react": "^16.4.2",
     "react-dom": "^16.4.2"
