--- src/systems/generic/implementations/react/blockquote/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@ecl/generic-react-component-blockquote",
-  "author": "European Commission",
-  "license": "EUPL-1.1",
-  "version": "2.0.0-alpha.0",
-  "description": "ECL Generic React Blockquote",
-  "main": "Blockquote.jsx",
-  "module": "Blockquote.jsx",
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ec-europa/europa-component-library.git"
-  },
-  "bugs": {
-    "url": "https://github.com/ec-europa/europa-component-library/issues"
-  },
-  "homepage": "https://github.com/ec-europa/europa-component-library",
-  "keywords": [
-    "ecl",
-    "europa-component-library",
-    "design-system",
-    "react"
-  ],
-  "devDependencies": {
-    "prop-types": "^15.6.2",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2"
-  },
-  "peerDependencies": {
-    "prop-types": "^15.6.2",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2"
-  }
-}
