--- src/systems/generic/implementations/react/text-area/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@ecl/generic-react-component-text-area",
-  "author": "European Commission",
-  "license": "EUPL-1.1",
-  "version": "2.0.0-alpha.0",
-  "description": "ECL Generic React Text Area",
-  "main": "TextArea.jsx",
-  "module": "TextArea.jsx",
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ec-europa/europa-component-library.git"
-  },
-  "bugs": {
-    "url": "https://github.com/ec-europa/europa-component-library/issues"
-  },
-  "homepage": "https://github.com/ec-europa/europa-component-library",
-  "keywords": [
-    "ecl",
-    "europa-component-library",
-    "design-system",
-    "react"
-  ],
-  "peerDependencies": {
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
-    "prop-types": "^15.6.2"
-  }
-}
