--- src/systems/generic/implementations/react/text-input/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@ecl/generic-react-component-text-input",
-  "author": "European Commission",
-  "license": "EUPL-1.1",
-  "version": "2.0.0-alpha.0",
-  "description": "ECL Generic React Text Input",
-  "main": "TextInput.jsx",
-  "module": "TextInput.jsx",
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ec-europa/europa-component-library.git"
-  },
-  "bugs": {
-    "url": "https://github.com/ec-europa/europa-component-library/issues"
-  },
-  "homepage": "https://github.com/ec-europa/europa-component-library",
-  "keywords": [
-    "ecl",
-    "europa-component-library",
-    "design-system",
-    "react"
-  ],
-  "dependencies": {
-    "@ecl/generic-resources": "^2.0.0-alpha.0"
-  },
-  "peerDependencies": {
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
-    "prop-types": "^15.6.2"
-  }
-}
